+++ CHANGELOG.md
@@ -7,6 +7,10 @@ We go to the next version after each release on the
 
 ## [Unreleased]
 
+### Changed
+
+- The clock menu position widget to `AdwSpinRow`.
+
 ## [28.0.0 Lear] - 2024-03-06
 
 ### Added